+++ tests/test_order_scope.py
@@ -13,7 +13,7 @@ except ImportError:
 
 @pytest.fixture(scope="module")
 def fixture_path(tmpdir_factory):
-    fixture_path = str(tmpdir_factory.mktemp("fixtures"))
+    fixture_path = str(tmpdir_factory.mktemp("order_scope"))
     testname = os.path.join(fixture_path, "test_classes.py")
     test_class_contents = """
 import pytest